pyproject.toml
[project]
name = "ocr-studio"
version = "1.0.0"
description = "Windows desktop OCR & translation tool powered by OpenAI GPT-4o"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.7.0",
    "httpx>=0.27.0",
    "PyMuPDF>=1.24.0",
    "Pillow>=10.0.0",
    "mss>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "pytest>=8.0",
]

[project.gui-scripts]
ocr-studio = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
